Trace events removed because they are not implemented properly.
Trace events must be in try block because some users do not have permission to post event log messages or there may be some other error like WMI is disabled or something that will cause an exception while trying to send the trace.  Furthermore, average users don't want to be notified of things like an update check failing in their event log (especially if the reason is that they told their firewall not to give access to the internet), so trace messages should only be enabled in debug builds if they are needed for debugging purposes or reported as a result of catastrophic errors not everyday errors that do not effect the functioning of the software.


git-svn-id: https://svn.code.sf.net/p/willowtree/code/trunk@102 32260eb2-1dd4-49be-8fe3-fb4c10400f30